Revert "Update Wine to 7.8, refresh patches, remove more dlls"
This commit seems to be causing lag in the game loop appearing as "Too
much latency to simulate (...) for projectile lag compensation on ...".
I suspect this may be caused by the user32/win32u changes in Wine, but
since we don't really need the Wine update, I'm going to revert it for
now until I have time to look into the issue.

This reverts commit 57bdcff2bd7f6c80191a8dfb4f2647cca3d21374.